Make the imported game read and build as C23
Every kept translation unit now compiles under -std=c23 -O2 -Werror with
the warning set this project uses for application code, plus the node's
portability gates (-Wvla, -Walloca, -Wshift-overflow=2,
-Wstrict-prototypes, -Wundef and the rest); 71 of 71 are clean. Each
file carries the repo header and a purpose line, same author, same
licence.

The fixes, all minimal and behaviour-preserving unless noted:

- _Static_assert over float constants (54 diagnostics). C23 6.7.11 wants
  an integer constant expression and allows a floating constant only as
  the immediate operand of a cast, so each spec bound now casts. Every
  constant involved is a whole number in the unit it is cast in; the two
  that are not, PLAYER_RADIUS and GROUND_DETECTION_RANGE, gained an exact
  _TENTHS spelling that the float is derived from. No bound moved.
- enemies.c: the seven (Color){...} compound literals in a static
  initializer became plain braced initializers - same values, and now a
  constant expression.
- specification_enforcement.h: "NULL == 0" is an address constant, never
  an integer constant expression. C 6.3.2.3 guarantees what it asserted,
  so it is a comment now.
- Seven switches over enums gained a "default: break;", which is what
  falling out of them already did.
- 35 unused parameters and 9 unused locals took C23 [[maybe_unused]].
- crash_protection.c: write(2) is warn_unused_result and a signal handler
  has nowhere to report a short write, so one crash_write() helper
  consumes the result once instead of at 13 call sites.
- ai_pilot_load(): the six unchecked freads now have to all return 1, and
  a truncated file is refused instead of flying on half-loaded weights.
- input.c: fabsf() on a short became abs(), which is what it meant.
- input_model.c: the axis floor, axis ceiling and L2 full-press values sit
  exactly on short's own limits, so testing them at runtime could never
  fail. Three static_asserts pin the constants to SHRT_MIN/SHRT_MAX
  instead: a spec change that moves one inside short's range now breaks
  the build rather than passing a check that cannot fire.
- gdb_proof_view.c: M_PI is not ISO C, so a guarded define replaces it.
- camera_controller.c: dropped a static prototype with no definition.

Two more files did not survive. gdb_proof_no_crash.c refuses to compile
unless the game's GDB proof pipeline has defined
PATH_SPEC_SYSTEM_INIT_CALLED_EXISTS, and this project does not run that
pipeline - defining the macro by hand would assert a proof nobody
performed. sky_combat_ultimate_main.c is a five-line main calling
sky_combat_ultimate(), whose only definition was in one of the files that
could not be imported.

+++ apps/skycombat/specifications/joystick_control_specs.h
@@ -1,5 +1,6 @@
-/* SPDX-FileCopyrightText: 2025 Rhett Creighton
- * SPDX-License-Identifier: Apache-2.0
+/* Copyright 2026 Rhett Creighton - Apache License 2.0
+ *
+ * Sky Combat: joystick range and latency specification.
  */
 
 #ifndef JOYSTICK_CONTROL_SPECS_H
@@ -74,10 +75,10 @@ typedef struct {
 } while(0)
 
 #define PROVE_CONTROL_RATES() do { \
-    _Static_assert(MAX_PITCH_RATE > 0 && MAX_PITCH_RATE <= 180.0f, "Invalid pitch rate"); \
-    _Static_assert(MAX_ROLL_RATE > 0 && MAX_ROLL_RATE <= 360.0f, "Invalid roll rate"); \
-    _Static_assert(MAX_YAW_RATE > 0 && MAX_YAW_RATE <= 180.0f, "Invalid yaw rate"); \
-    _Static_assert(MAX_THROTTLE_RATE > 0, "Invalid throttle rate"); \
+    _Static_assert((int)MAX_PITCH_RATE > 0 && (int)MAX_PITCH_RATE <= 180, "Invalid pitch rate"); \
+    _Static_assert((int)MAX_ROLL_RATE > 0 && (int)MAX_ROLL_RATE <= 360, "Invalid roll rate"); \
+    _Static_assert((int)MAX_YAW_RATE > 0 && (int)MAX_YAW_RATE <= 180, "Invalid yaw rate"); \
+    _Static_assert((int)MAX_THROTTLE_RATE > 0, "Invalid throttle rate"); \
 } while(0)
 
 #define PROVE_RESPONSE_TIME() do { \
